Record metrics for `ResourceNode` events (#6557)
Motivation:

This is the last PR of the series for #6464 .

The changeset attempts to add observability into `ResourceNode`s in the
`xds` module.
More specifically, `onChanged` records the current revision of the node,
and `onError` records how many errors were recorded for a specific
`ResourceNode`.
The `resourceName` and `type` is used as the default differentiator. If
there are duplicates, the expectation is that only one of the nodes will
be recorded.

Currently, child `ResourceNode`s are created before they are closed.

https://github.com/line/armeria/blob/adbf8755c8b2ae84f38ab1da72cf23e98f1c0f86/xds/src/main/java/com/linecorp/armeria/xds/ClusterResourceNode.java#L47-L51

This is in order to ensure that `XdsLoadBalancer`'s state is propagated
across resource updates. (i.e. ClusterA updates doesn't result in
calling `ClusterResourceNode.close`, which in turn allows ramping-up
behavior so survive updates).

On the other hand, metrics that are bound to a `ResourceNode` must be
closed before a new `ResourceNode` is created.
Otherwise, a cached `Gauge` may be returned which doesn't reflect the
state of the new `ResourceNode`.

For this, I propose a new `preClose` lifecycle method is introduced.
This is called before the child node's `close` is called and will be
used to unbind metrics before creating a new node.

Modifications:

- Added `AbstractNodeSnapshotWatcher` which abstracts logic for handling
`closed` early return and calling `preClose` to child nodes
- Added `ResourceNodeMeterBinder` which binds metrics for a
`ResourceNode`
- `AbstractResourceNode` binds to a `ResourceNodeMeterBinder` and calls
callback methods for i
- Each `ResourceNode` implementation calls `preClose` before updating
the child node. Also, `preClose` calls are propagated to child nodes for
cases where only the top-level node is updated.

Result:

- `ResourceNode` specific metrics are exported

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->